Refactor multiplication, subtraction, and addition tests to use double instead of int for better precision

+++ src/test/java/com/calculator/components/calculator/MultiplyTest.java
@@ -18,70 +18,70 @@ private Calculator calc;
     @Test
     public void testPerkalianBilanganPositif() {
         // (1) setup data
-        int AngkaPertama = 100;
-        int AngkaKedua = 100;
+        double AngkaPertama = 100;
+        double AngkaKedua = 100;
         
         // (2) panggil method yang akan di test
-        double result = calc.multiply((double)AngkaPertama, (double)AngkaKedua);
+        double result = calc.multiply(AngkaPertama, AngkaKedua);
         
         // (3) bandingkan hasilnya dengan yang diharapkan
-        int expected = 10000;
+        double expected = 10000;
         assertEquals(expected, result, 0.01);
     }
 
     @Test
     public void testPerkalianBilanganNegatif() {
         // (1) setup data
-        int AngkaPertama = -50;
-        int AngkaKedua = -100;
+        double AngkaPertama = -50;
+        double AngkaKedua = -100;
         
         // (2) panggil method yang akan di test
-        double result = calc.multiply((double)AngkaPertama, (double)AngkaKedua);
+        double result = calc.multiply(AngkaPertama, AngkaKedua);
         
         // (3) bandingkan hasilnya dengan yang diharapkan
-        int expected = 5000;
+        double expected = 5000;
         assertEquals(expected, result, 0.01);
     }
 
     @Test
     public void testPerkalianBilanganPositifNegatif() {
         // (1) setup data
-        int AngkaPertama = 100;
-        int AngkaKedua = -5;
+        double AngkaPertama = 100;
+        double AngkaKedua = -5;
         
         // (2) panggil method yang akan di test
-        double result = calc.multiply((double)AngkaPertama, (double)AngkaKedua);
+        double result = calc.multiply(AngkaPertama, AngkaKedua);
         
         // (3) bandingkan hasilnya dengan yang diharapkan
-        int expected = -500;
+        double expected = -500;
         assertEquals(expected, result, 0.01);
     }
 
     @Test
     public void testPerkalianBilanganPositifNol() {
         // (1) setup data
-        int AngkaPertama = 100;
-        int AngkaKedua = 0;
+        double AngkaPertama = 100;
+        double AngkaKedua = 0;
         
         // (2) panggil method yang akan di test
-        double result = calc.multiply((double)AngkaPertama, (double)AngkaKedua);
+        double result = calc.multiply(AngkaPertama, AngkaKedua);
         
         // (3) bandingkan hasilnya dengan yang diharapkan
-        int expected = 0;
+        double expected = 0;
         assertEquals(expected, result, 0.01);
     }
 
     @Test
     public void testPerkalianBilanganNol() {
         // (1) setup data
-        int AngkaPertama = 0;
-        int AngkaKedua = 0;
+        double AngkaPertama = 0;
+        double AngkaKedua = 0;
         
         // (2) panggil method yang akan di test
-        double result = calc.multiply((double)AngkaPertama, (double)AngkaKedua);
+        double result = calc.multiply(AngkaPertama, AngkaKedua);
         
         // (3) bandingkan hasilnya dengan yang diharapkan
-        int expected = 0;
+        double expected = 0;
         assertEquals(expected, result, 0.01);
     }
 }
