+++ src/test/java/com/calculator/components/calculator/SubtractTest.java
@@ -18,56 +18,56 @@ public class SubtractTest {
     @Test
     public void testPenguranganBilanganPositif() {
         // (1) setup data
-        int AngkaPertama = 100;
-        int AngkaKedua = 100;
+        double AngkaPertama = 100;
+        double AngkaKedua = 100;
         
         // (2) panggil method yang akan di test
-        double result = calc.subtract((double)AngkaPertama, (double)AngkaKedua);
+        double result = calc.subtract(AngkaPertama, AngkaKedua);
         
         // (3) bandingkan hasilnya dengan yang diharapkan
-        int expected = 0;
+        double expected = 0;
         assertEquals(expected, result, 0.01);
     }
 
     @Test
     public void testPenguranganBilanganNegatif() {
         // (1) setup data
-        int AngkaPertama = -100;
-        int AngkaKedua = -200;
+        double AngkaPertama = -100;
+        double AngkaKedua = -200;
         
         // (2) panggil method yang akan di test
-        double result = calc.subtract((double)AngkaPertama, (double)AngkaKedua);
+        double result = calc.subtract(AngkaPertama, AngkaKedua);
         
         // (3) bandingkan hasilnya dengan yang diharapkan
-        int expected = 100;
+        double expected = 100;
         assertEquals(expected, result, 0.01);
     }
 
     @Test
     public void testPenguranganBilanganPositifNegatif() {
         // (1) setup data
-        int AngkaPertama = 100;
-        int AngkaKedua = -200;
+        double AngkaPertama = 100;
+        double AngkaKedua = -200;
         
         // (2) panggil method yang akan di test
-        double result = calc.subtract((double)AngkaPertama, (double)AngkaKedua);
+        double result = calc.subtract(AngkaPertama, AngkaKedua);
         
         // (3) bandingkan hasilnya dengan yang diharapkan
-        int expected = 300;
+        double expected = 300;
         assertEquals(expected, result, 0.01);
     }
 
     @Test
     public void testPenguranganBilanganNol() {
         // (1) setup data
-        int AngkaPertama = 0;
-        int AngkaKedua = 0;
+        double AngkaPertama = 0;
+        double AngkaKedua = 0;
         
         // (2) panggil method yang akan di test
-        double result = calc.subtract((double)AngkaPertama, (double)AngkaKedua);
+        double result = calc.subtract(AngkaPertama, AngkaKedua);
         
         // (3) bandingkan hasilnya dengan yang diharapkan
-        int expected = 0;
+        double expected = 0;
         assertEquals(expected, result, 0.01);
     }
